python/requirements.txt
# Little Lemon Database Capstone Project
# Python Dependencies
# Fecha: 10 de Julio, 2025

# Database connectivity
mysql-connector-python==8.0.33

# Data analysis and manipulation
pandas==2.0.3
numpy==1.24.3

# Data visualization
matplotlib==3.7.1
seaborn==0.12.2

# Jupyter notebook for analysis
jupyter==1.0.0
notebook==6.5.4

# Web framework (if needed for future web interface)
flask==2.3.2

# Configuration management
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Documentation
sphinx==7.1.1

# Date and time utilities
python-dateutil==2.8.2

# Logging and monitoring
colorlog==6.7.0

# Excel export (for reports)
openpyxl==3.1.2

# CSV handling
csv-diff==1.0

# JSON handling
jsonschema==4.19.0

# HTTP requests (for APIs)
requests==2.31.0

# Development tools
black==23.7.0
flake8==6.0.0
mypy==1.4.1
